api/requirements.txt
aniso8601==4.0.1
argcomplete==1.9.4
astroid==1.6.5
backports.functools-lru-cache==1.5
Click==7.0
configparser==3.5.0
enum34==1.1.6
Flask==1.0.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.14.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
futures==3.2.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
lxml==4.2.5
MarkupSafe==1.1.0
mccabe==0.6.1
passlib==1.7.1
pkg-resources==0.0.0
pycurl==7.43.0.2
PyJWT==1.4.2
pylint==1.9.3
pytz==2018.7
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.2.15
webdavclient==1.0.8
Werkzeug==0.14.1
wrapt==1.10.11
